modifiquei as questões 1 para 2 pois estavam erradas

+++ 2026.1/analise_algoritmo_1/questao2_a.cpp
@@ -3,9 +3,9 @@
 
 
 int main(){
-    int n;
+    long n;
     std::cin >> n;
-    double t = pow(2,n)/(n*n) - n + 10;
+    long t = pow(2,n)/(n*n) - n + 10;
     std::cout << t << std::endl;
     return 0;
 }
